+++ tests/Codec/Picture/ExtraSpec.hs
@@ -52,6 +52,9 @@ spec = do
   describe "flipVertically"    flipVerticallySpec
   describe "rotateLeft90Spec"  rotateLeft90Spec
   describe "rotateRight90Spec" rotateRight90Spec
+  describe "rotate180"         rotate180Spec
+  describe "beside"            besideSpec
+  describe "below"             belowSpec
 
 scaleBilinearSpec :: Spec
 scaleBilinearSpec = do
@@ -110,6 +113,30 @@ rotateRight90Spec =
       checkWithFiles rotateRight90
         "data-examples/lenna.png"
         "data-examples/lenna-right-rotated.png"
+
+rotate180Spec :: Spec
+rotate180Spec =
+  context "when we rotate by 180°" $
+    it "produces correct image" $
+      checkWithFiles rotate180
+        "data-examples/lenna.png"
+        "data-examples/lenna-180-rotated.png"
+
+besideSpec :: Spec
+besideSpec =
+  context "when we place images beside each other" $
+    it "produces correct image" $
+      checkWithFiles (\x -> beside [x,x])
+        "data-examples/lenna.png"
+        "data-examples/lenna-beside.png"
+
+belowSpec :: Spec
+belowSpec =
+  context "when we place images below each other" $
+    it "produces correct image" $
+      checkWithFiles (\x -> below [x,x])
+        "data-examples/lenna.png"
+        "data-examples/lenna-below.png"
 
 -- | Run given transforming function on image loaded from one file and
 -- compare resulting image with contents of another file.